// pbweberogwcl/src/app/registro-controlli/commons/models/beneficiario.ts
/*
 * Copyright Regione Piemonte - 2024 
 * SPDX-License-Identifier: EUPL-1.2 
*/

export class Beneficiario {
  id_soggetto: number;
  id_progetto: number;
  codiceFiscale: string;
  cognome: string;
  nome: string;
  descrizione: string;
  id: number;
  idFormaGiuridica: number;
  idDimensioneImpresa: number;
}